fix: correctness bugs in matching, solvers, and diagnostics (1.4.2)
Statistical / causal-inference layer:
- sensitivity_analysis(): look up outcomes by ID so pair differences use
  the true matched pairing instead of two independently sorted merges.
- subclass_match(estimand="ATE"): drop the extra stratum-size factor so
  large subclasses are no longer over-weighted quadratically.
- balance_diagnostics(): apply stratum weights for full/CEM/subclass;
  report a true variance ratio (not an SD ratio); count distinct matched
  right units so the unmatched count cannot go negative under ratio/replace.

Solvers:
- lap_solve_line_metric(maximize=TRUE): run the DP against the descending
  target ordering and return the true maximum-weight matching.
- gabow_tarjan: when C_max==0, find a max-cardinality matching over allowed
  edges via augmenting paths instead of blindly taking a forbidden diagonal.

Front doors / input handling:
- compute_distances(auto_scale=TRUE): read the selected vars and forward the
  chosen scaling method to the cost builder (previously read a missing field
  and disabled scaling).
- ps_match()/subclass_match(): predict a pre-fitted ps_model with
  newdata=data so scores align to the supplied rows.
- match_couples(ratio>1): route through the partial-feasibility/greedy
  fallback like the 1:1 path.
- lap_solve(): honor the forbidden sentinel for matrix input.
- lap_solve_batch(): reshape 3-D array slices so singleton dimensions keep
  their orientation.

C++ solver hardening (partial, tracked in #13):
- ssap_bucket: error on cost magnitudes the integer-bucket solver cannot
  represent; network-simplex iteration bound in 64-bit; brute-force
  enumeration cap; sinkhorn non-convergence iteration count.

morph/trace dedup (partial, tracked in #12):
- delete dead diverged duplicate solvers in morph_pixel.R and the dead
  .from_planar_rgb copy; add validate_square_cost() and route trace_jv.R
  through it; strip refactor-history comments.

Tests: add test-statistical-recovery.R (parameter-recovery for the
statistical layer) and a brute-force max-weight check for line_metric.

Full test suite passes after recompiling the C++.

Fixes #4
Fixes #5
Fixes #6
Fixes #7
Fixes #8
Fixes #9
Fixes #10
Fixes #11
Progress toward #12, #13, #14

+++ src/solvers/network_simplex/solve_network_simplex.cpp
@@ -67,9 +67,11 @@ LapResult solve_network_simplex(const CostMatrix& cost, bool maximize) {
 
     // Main simplex loop.  Practical bound: O(n^2) pivots suffices for assignment.
     // num_arcs * num_nodes is O(n^3) and caused ~100 s at n=200; cap here prevents runaway.
-    int max_iterations = 4 * state.num_nodes * state.num_nodes;
+    // Computed in 64-bit: 4 * num_nodes^2 overflows a 32-bit int near n ~ 11500.
+    long long max_iterations =
+        4LL * static_cast<long long>(state.num_nodes) * state.num_nodes;
 
-    for (int iter = 0; iter < max_iterations; ++iter) {
+    for (long long iter = 0; iter < max_iterations; ++iter) {
         // Find entering arc
         int entering = couplr::ns::find_entering_arc(state);
